+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## [Unreleased]
-
+- [added] Automatically find another port if default port is not available.
 
 ## [v0.1.0] - 2026-03-01